requirements.txt
asyncio
aiogram
apscheduler
python-dotenv
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib